requirements.txt
# Core Frameworks
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
llama-index>=0.9.25
llama-index-core>=0.10.0
llama-index-agent-openai>=0.1.4
llama-index-vector-stores-qdrant>=0.1.2
llama-index-graph-stores-neo4j>=0.1.2
llama-index-embeddings-openai>=0.1.5
llama-index-embeddings-cohere>=0.1.2
llama-index-readers-web>=0.1.3
llama-index-cli>=0.1.2

# LLM Clients
openai>=1.13.0
anthropic>=0.7.0
cohere>=4.46

# Memory & Vector Storage
qdrant-client>=1.7.0
# mem0-client>=0.1.0  # Install from github when available

# Graph Database & Knowledge Representation
neo4j>=5.14.0
# graphiti>=0.1.0  # Install from github when available
rdflib>=7.0.0
owlready2>=0.45

# Structured Data
supabase>=2.3.0
postgrest>=0.13.0
psycopg2-binary>=2.9.9

# Embeddings
fastembed>=0.2.0

# MCP Server
fastmcp>=0.1.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.5.0

# UI Framework
chainlit>=1.0.0

# Web Scraping & Ingestion
crawler4ai>=0.0.5
firecrawl>=0.1.0
scrapy>=2.11.0
playwright>=1.40.0
beautifulsoup4>=4.12.3
requests>=2.31.0
aiohttp>=3.9.0
feedparser>=6.0.10
python-imap>=0.0.2

# AWS Integration
boto3>=1.34.0
s3fs>=2023.10.0

# Monitoring & Observability
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Security & Utilities
python-dotenv>=1.0.0
pyjwt>=2.8.0
cryptography>=42.0.0
tenacity>=8.2.3
aiofiles>=23.2.1
pyyaml>=6.0.1
tqdm>=4.66.1
python-dateutil>=2.8.2

# Development Dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
ruff>=0.1.5
mypy>=1.7.0
pre-commit>=3.5.0
uv>=0.1.11